SkJpegMultiPicture: Add encoding support (part 1)
Change SkJpegParseMultiPicture to be a static member function
SkJpegMultiPictureParameters::Make.

Add the function SkJpegMultiPictureParameters::serialize. Given
SkJpegMultiPictureParameters, it produces a Jpeg segment
representation of those parameters. This function is in effect the
inverse of SkJpegParseMultiPicture.

Fix a bug in SkJpegParseMultiPicture in which the value for kMPEntryTag
was ignored. The value should be interpreted as the offset for the
array of MP entries.

Add tests!

Bug: skia:14031
Change-Id: Ie50cf2d7230d80bae01a307b797d8113e69bedf2
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/642776
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Christopher Cameron <ccameron@google.com>