SkGainmapShader: Support alpha and red single-channel images
The gainmap shader had an assumption built in that the gainmap image,
when sampled, would populate the r,g,b channels. This is not the case
for kAlpha_8_SkColorType images (which is a common representation of
single-channel images).

Update the shader to take this into account.

Also fix a bug where my printf-debugging ended up landing (eep).

Bug: skia:14031
Change-Id: I7ada1be2ae2d3f9d11d432abb5499a83015f857a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/643577
Reviewed-by: John Reck <jreck@google.com>
Commit-Queue: Christopher Cameron <ccameron@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>