SkJpegCodec: Create MPF component
Create a jpeg_mpf component, which will be used for both encode
and decode.

The files in this component depended on constants that were in the
header file SkJpegPriv.h, which also includes libjpeg, which is
not required by this component. Move the constants that were in
SkJpegPriv.h to SkJpegConstants.h.

Bug: skia:14031
Change-Id: I016f9b042664e40531d5c4075770d1a420b7cfa3
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/644017
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Christopher Cameron <ccameron@google.com>