Fix GLSL codegen of empty trailing switch-case blocks.
The GLSL spec insists that the last case in a switch statement
must have an associated statement. In practice, the Apple GLSL
compiler crashes if that statement is a no-op, such as a semicolon
or an empty brace pair. (This is filed as FB11992149.) It also
crashes if we put two `break` statements in a row. To work around
this while honoring the rules of the standard, we inject an extra
break if and only if the last switch-case block is empty.

Bug: skia:14126
Change-Id: I9272aa9f4fcc2a82db4cb85c5b5831306e4e3aa0
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/644498
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: Arman Uguray <armansito@google.com>
Reviewed-by: Arman Uguray <armansito@google.com>