SkJpegMultiPictureParameters: Fix version verification
Unlike other tag values, the MP parameters version should be treated
as a sequence of bytes, and not as an integer.

Prior to this patch, we would apply bytes swapping to the MP
parameters version. The MP parameters can configure their endian-ness,
and this happened to work when the configured endian-ness was
big-endian.

This fails when the file parameters configure themselves as
small-endian.

Change the macro DEFINE_AND_READ_UINT to also preserve the raw
bytes, and use them for the version comparison.

Add lots of tests for parsing MP parameters.

Bug: skia:14031
Change-Id: Ic833d74de166e3d1b9108b415407fddcb74f06d1
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/644636
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Christopher Cameron <ccameron@google.com>