SkJpegXmp: Do not hardcode namespaces
Prior to this patch, we hard-coded the namespaces for the interfaces
we were using.

For instance, we hard-coded
  xmlns:GContainer="http://ns.google.com/photos/1.0/container/"

Any string can be used in place of GContainer. A common example is
  xmlns:Container="http://ns.google.com/photos/1.0/container/"

Change SkJpegXmp::findNamespaceUriMatch to be named findUriNamespaces,
and make it populate the namespace (instead of asserting that it has
a particular value).

Bug: skia:14031
Change-Id: Ie398a619a79ccb421e17152059abbac3efa5175d
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/644616
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Christopher Cameron <ccameron@google.com>