SkJpegCodec: Merge JpegR and HDRGM
An HDRGM image is a MultiPicture image, where one of the auxiliary
images identifies itself as a gainmap by having HDRGM XMP metadata.

A JpegR file now refers to an HDRGM image which also contains
GContainer XMP metadata on the primary image. The GContainer XMP
metadata also identifies where the gainmap image is in the file
(namely, it is where the item with Semantic "RecoveryMap" is).

Change SkJpegXmp::getGainmapInfoJpegR to now be the function
getContainerGainmapLocation, which returns only the location of the
item with Semantic "RecoveryMap". It no longer populates
SkGainmapInfo, because that information is now within the gainmap
image itself.

JpegR no longer has its own gainmap rendering parameters (e.g, the
rangeScalingFactor parameter) in the primary image, and instead is
now much more similar to the MPF path. Remove the decoding code in
SkJpegGetJpegRGainmap, and merge the gainmap path with the MPF path.
Move the functions out of SkJpegGainmap.cpp/h and into
SkJpegCodec.cpp/h.

Also remove the encoding code that would convert the more general
HDRGM parameters to JpegR. Leave SkJpegGainmapEncoder::EncodeJpegR
present, but have it call HDRGM.

Bug: skia:14031
Change-Id: I76d4c1ae191af6e6764c752a0416d81977501acc
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/644637
Commit-Queue: Christopher Cameron <ccameron@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>