Add RP builder op for reenable_loop_mask from the stack.
Previously, we used reenable_loop_mask for continue statements,
which store their mask in a slot. For switch statements, the
default-mask is stored on the stack, so we now need a
"pop-and-reenable" variation of the op which takes its input from
the stack.

Change-Id: Ie46980eec92de93156ebc6dbbb4fffbd0cf2fc6c
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/644656
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>