Implement switch statement in RP codegen.
We have a dedicated case_op now; this manages the case-value checking
as well as keeping track of lanes which never matched anything. Lanes
which never matched any case are eligible for the "default" case. The
loop mask is used to control execution.

This causes 8 more tests to pass.

Change-Id: I269fd9ab9424f6cf9feabb5a4c958a4fb1f24e80
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/644497
Reviewed-by: Brian Osman <brianosman@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>