Eliminate no-op branches via peephole optimization.
When emitting a label (a branch target), the Builder now checks to see
if the prior instruction was a branch to that label. If so, the branch
was unnecessary, so we can remove it. The label itself is still
written, since it's technically valid for more than one branch to
target the same label.

This gives us better code generation for consecutive cases in a
switch.

Change-Id: I828fcf12b78e8eb5ca50dd11dc281393ad4a74f6
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/644657
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>