Remove SlotMap from LValues.
Now, the lvalues are directly responsible for pushing and storing.
This adds flexibility because it introduces the possibility that an
LValue can be sourced from a temporary stack.

Additionally, LValue::push now returns a bool, in order to handle
the (soon to exist) case of an unsupported temporary expression.

Change-Id: I29315d7db5a5baa572f9c6be3a033e64b9617dc3
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/645056
Commit-Queue: Brian Osman <brianosman@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>