SkJpegEncoder: Ensure MPF segment be written last
From SkJpegEncoder and SkJpegEncoderMgr and skjpeg_destination_mgr,
remove the "suffix" parameter. It is no longer used.

Instead of listing a bunch of segments as parameters to the private
SkJpegEncoder::Make, add a SkJpegEncoder::OptionsPrivate for options
that aren't quite ready to add to the public APIs yet.

In SkJpegEncoder::OptionsPrivate, include the XMP metadata, and the
full MPF segment. The XMP metadata probably makes sense to make into
a public option soon (while the MPF segment is a bit more complicated).

Ensure that the MPF segment be written last (after XMP, and ICC
profile), to ensure that edits to those segments do not break the
MPF offsets.

Bug: skia:14031
Change-Id: I2ba66b1d34cd4f2a96bd54ac8a6a68695f10fbb2
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/645076
Commit-Queue: Christopher Cameron <ccameron@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>