[sksl][metal] Add test case that demonstrates nested out-param invocation bugs
Added at test case that demonstrates the issues described in skia:14130.
This test also demonstrates a bug in Metal coden that causes an
out-param helper definition to be emitted in the middle of an argument
expression when an out-param function is invoked with an argument that
itself contains another out-param function invocation (e.g. in an array
index expression).

Bug: skia:14130

Change-Id: Ifa8357e95fbc0e0b679e90ee665fa1cfc4eec60e
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/645676
Reviewed-by: John Stiles <johnstiles@google.com>
Commit-Queue: Arman Uguray <armansito@google.com>