[sksl][metal] Fix invalid codegen with nested out-param invocations
The Metal codegen generated incorrect code when an out-param function is
invoked with an argument that itself contains another out-param function
invocation (e.g. in an array index expression). The bug caused the inner
expression to emit an out-param helper function definition in the middle
of the argument expression inside the outer out-param helper.

Bug: skia:14130

Change-Id: Iaad8f25084b44eebb0b887ebeb3f12f6898147d3
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/645677
Commit-Queue: Arman Uguray <armansito@google.com>
Reviewed-by: John Stiles <johnstiles@google.com>