SkJpegXmp: Support more of xmp
Update find_uri_namespaces to allow namespaces to be specified in
both the rdf:RDF and rdf:Description nodes.

Add the function get_attr (and related helper functions) that allow
parameters to be specified either as attributes or as separate fields.
E.g, treating as equivalent both
    <rdf:Description rdf:about="">
        <hdrgm:Version>1.0</hdrgm:Version>
    </rdf:Description>
and
    <rdf:Description rdf:about="" hdrgm:Version="1.0"/>

Add support for typed nodes using RDF. E.g, treating as equivalent
both
    <Container:Item>
        <Item:Semantic>Primary</Item:Semantic>
    </Container:Item>
and
    <rdf:value rdf:parseType="Resource">
        <Item:Semantic>Primary</Item:Semantic>
    </rdf:value>
    <rdf:type rdf:resource="Item"/>

Add TODOs for further support of less common syntax. E.g, in the above
syntax "Item" could instead be the full URI for "Item".

We will likely want to switch to a real XMP library for this parsing at
some point in the future.

Bug: skia:14031
Change-Id: Ic7c933656405fdce88d45445b0ca6926e981a1ab
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/646296
Commit-Queue: Christopher Cameron <ccameron@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>