SkAndroidCodec: Remove "wide" gamut heuristic, use actual gamut
The logic here was faulty (the cutoff was low enough that we considered
sRGB images as "wide"). Further, there's no reason to be binning images
to sRGB or P3 here -- let's just use the actual gamut if possible.

If we can't, just fall back to sRGB, like we were doing before, anyway.
The test change is because we're now calling MakeRGB for this image, and
despite the name, it has a gamut that is VERY close to sRGB, so we
actually snap the resulting color space (for performance reasons).

Bug: skia:14137
Change-Id: I902481604d62b3e809eadf73a6bb9d50f5710f16
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/647039
Reviewed-by: Leon Scroggins <scroggo@google.com>
Reviewed-by: John Reck <jreck@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>