[graphite] Add EdgeAAQuad to Geometry union
Defines EdgeAAQuad as 4 xs, 4 ys, an AA flags bitmask, and a boolean to
remember if it was defined by a rectangle or arbitrary points. I opted
to put this in Geometry instead of Shape because of this extra metadata
that goes beyond a pure "shape".

Updates the logic in AnalyticRRectRenderStep to upload the EdgeAAQuad
values, although the vertex shader needs to be updated to handle the
edge flags (the hard-coded vertex normals are only valid for rects,
so the further the quad deviates from that, the weird the outsets get).

Implements the solid-color drawEdgeAAQuad in Device. drawEdgeAAImageSet
will be implemented in another CL.


Bug: b/241403145
Change-Id: I889bc7497bd97a0e9f5f036d133c5863ce1fc13b
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/645038
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>