Make sk_realloc_throw always free buffer if size == 0
realloc has implementation-dependent behavior [1][2] if size
is 0. We would like the versions in sk_realloc to be
implementation-independent for consistent behavior,
so this makes realloc always free the buffer if the
size was 0 and return nullptr.

This also adds some simple tests for our implementations
in SkMalloc, since I could not find any existing, dedicated
tests.

[1] https://en.cppreference.com/w/cpp/memory/c/realloc
[2] https://www.open-std.org/jtc1/sc22/wg14/www/docs/n2396.htm#dr_400

Change-Id: I160b0ed83134061935ff68f806e4019eb4e95ba0
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/647456
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Herb Derby <herb@google.com>
Reviewed-by: John Stiles <johnstiles@google.com>