Add Raster Pipeline op for indirect loads.
This op will load `slots` values from *(src + indirectOffset) and
copy them to *dst. `indirectOffset` points to a vector, so we use
a gather op to load the data. We clamp `indirectOffset` to a
passed-in `indirectLimit` to avoid indexing past the end of the
source buffer; if a load goes out of bounds, we will end up reading
data from the end of the source buffer. (This is fine; we don't
make any promises about what you get back if you over-index a
buffer.)

This is the first in a family of similar indirect-load ops that
we will undoubtedly need, such as:
- Load uniforms from *(src + indirectOffset) and write to *dst
  (the uniform data does not have SIMD lanes so the code will need
  to be slightly different)
- Load from *src, write to *(dst + indirectOffset)
  (indirection applied to the destination, not the source)
- Load from *src, swizzled-write to *(dst + indirectOffset)

Change-Id: I7777534b5b38c8f503089000569d0daa3c653965
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/647839
Reviewed-by: Herb Derby <herb@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>