Change copy_from_indirect_masked to ..._unmasked.
After submitting this, I realized it will be used as a gadget to
push values onto the stack. The stack doesn't need write masking,
so this doesn't need to honor the write mask.

The coming-soon copy_to_indirect op will be used to implement
popping from stack to slots, and this one will need to be masked.
(It will also be different in other ways; it's a scatter op, not a
gather op.)

Sorry for the churn!

Change-Id: Id195fb9839c2a3c9e816c892feb7b0c494ae323b
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/648298
Reviewed-by: Herb Derby <herb@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>