Experiment: enable SK_ENABLE_SKSL_IN_RASTER_PIPELINE.
SkRP is not fully functional yet, but we will fall back to SkVM
in cases that SkRP does not yet support.

This CL is a trial to see what will happen when we enable SkRP
on a broad scale. The expectation is that several golden images
will change slightly, and we will likely see performance deltas
(good and/or bad).

The change will most likely be reverted by Monday or sooner.

Change-Id: I7a92ab2cef322c157073681f63ab7b077ea337a1
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/637096
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>