Fix stack_rewind op errors when used with extend().
Previously, extend() was oblivious to the rewind context, and didn't
create a fRewindCtx if one was needed; this meant that the pipeline
wouldn't get a stack_checkpoint at all, and the first rewind would
halt the program (and perhaps also scribble over some memory first).
Now, we create a rewind context when one is needed, and update
rewind ops to use it.

Change-Id: Ia64f1468780ea5e9f999b35011f9859be37e3cec
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/648961
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>