SkJpegEncoder: Allow YUVAPixmaps as input
Add the functions SkJpegEncoder::Encode and SkJpegEncoder::Make which
take an SkYUVAPixmaps and SkColorSpace instead of an SkPixmap.

The core of this support is a yuva_copy_row function which re-packs
the YUV planes as a single plane with Y,U,V triples. This is a terrible
and inefficient scheme, but it can be replaced with more efficient
versions without changing the interface.

Also support encoding kAlpha_8_SkColorType and  kR8_unorm_SkColorType
SkPixmaps as JCS_GRAYSCALE, since this comes up often.

Bug: skia:14031
Change-Id: Ia4ad02f92b900573fb802d67062892f1aedb6ab9
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/648216
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Christopher Cameron <ccameron@google.com>