Remove SK_HAS_ANDROID_CODEC guard
It was only used in one place and was on by default in nearly every
build. Let's try removing the guard.

This requires breaking a dependency knot between SkCodec
and SkAndroidCodec. If a frame depends on frame N-1,
SkCodec would need to have SkAndroidCodec decode frame N-1, which
in turn requires a call to handleFrameIndex at N-1.

Instead of having SkCodec know about SkAndroidCodec, I turned
the call into a callback function, which *only* SkAndroidCodec sets.

This logic confused me greatly until I realized that this was
(complicated) recursion and then what the base case was (finding
a frame that does not need an earlier frame). I added some comments
to hopefully make that connection easier for future readers of
the code.

I also learned that if you want to be able to pass in a lambda
as a function parameter, you cannot use a function pointer
but must take in a std::function.

See also the discussion on https://skia-review.googlesource.com/c/skia/+/333225?tab=comments

Change-Id: I71c2e393d1a81b782f04433750134f32ea7dd2be
Bug: skia:13983
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/645436
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Owners-Override: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Leon Scroggins <scroggo@google.com>