Only use function return slots when necessary.
If a function only has a single return statement, we can just leave
the value from the return statement on the stack instead of popping
it to slots and then pushing it back to the stack.

This has a very large impact on the test corpus because the majority
of our tests have a single return from main(). Then again, this is
true of most real-world shaders too.

Change-Id: I42ed9e4e134054835716db3e060e3d2fdc68d0ab
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/649138
Reviewed-by: Arman Uguray <armansito@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>