Fix divide by zero in GrQuad::point().
In ClipStack's shape_contains_rect() we check to see if a transformed
Rect intersects the W=0 plane, but then we outset it, which may push
the Rect through the W=0 plane. In the test case it intersects the plane
exactly, which is not caught by the check. Doing the check after the
outset should address this.

Bug: oss-fuzz:53990
Change-Id: I5db56ee2ae8681772bce7ba258a709a4fdc9e911
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/650277
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>