Add RP op branch_if_all_lanes_active.
As the name suggests, this provides a way for us to branch when
the execution mask is fully enabled. In particular, this can let
us skip over dead code when `break` or `continue` is used in
uniform control-flow.

Change-Id: I2ea2b421eead1e3815c294848305eeb5d8fd7329
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/650278
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>