Add 'break target' labels to looping constructs.
These labels are added anywhere that `break` or `continue` would be
viable. The labels are not used anywhere yet; this will happen in a
followup.

Labels are compiled out when the final SkRP program is emitted, so
these programs should be roughly identical to their previous state
during execution.

Change-Id: Ib622376dff8a5bab2e41f57502727aadb373a362
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/649990
Reviewed-by: Herb Derby <herb@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>