Add callback to notify when GrDirectContext is being deleted.
It can often be hard for clients to track which objects are keeping a
GrDirectContext alive even though they dropped their direct refs to
it. For example SkImage and SkSurface may be keeping refs to the
GrDirectContext.

This callback allows the client to know when the GrDirectContext is
actually going away so they know it is safe to destroy the backing
GPU context that needs to outlive the GrDircetContext.

Bug: b/266626090
Change-Id: I174076d1b322e467bf95d5474069f5c0568e144c
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/650782
Reviewed-by: Robert Phillips <robertphillips@google.com>
Reviewed-by: John Reck <jreck@google.com>
Commit-Queue: Greg Daniel <egdaniel@google.com>