Experiment with simpler test API
The serializeTypefaceForTest API on SkStrikeServer has a lot
of machinery that is unused by the remote cache system.

Experiment with a simpler API for tracking typefaces through
the tests. Just remember the unique ID for the server typeface
and use that to look up the client typeface ID after
deserializing the strike changes.

Remove the test SkRemoteGlyphCache_TypefaceSerialization because
it exercises a part of the API that should never be used by
clients. Move the important part to
SkRemoteGlyphCache_StrikeSerialization.

Bug: skia:14163
Change-Id: Ic7858156b9d632a8a7ae22f075673f38a7a508ff
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/653279
Commit-Queue: Herb Derby <herb@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>