Add Skia buffer based serialization to SkChromeRemoteGlyphCache
Add the top level of the Skia based serialization code. This
is tested by the existing remote glyph cache unit tests.

The two serialization systems, the std::vector and the skia based,
are controlled by the flag:
  SK_SUPPORT_LEGACY_STRIKE_SERIALIZATION

This CL keeps the old system as the default, but his will change
soon.

Add a bunch of #if defined to keep the compiler happy.

Bug: skia:14163
Change-Id: I1f2b709b713ed2addba2b7bf025f84a8629275f5
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/656238
Commit-Queue: Herb Derby <herb@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>