Restore Android-specific code in Ganesh's Perlin noise code
We need this code for Android for the following gpus:
  Tegra3, PowerVRGE8320 (Wembley), MaliG76, and Adreno308
However, if we always turn it on, we get failures on:
  iOS/PowerVRGT7600 and iOS/PowerVRGT7800
Boosting the pinning code to float (from half) precision fixes the issues on the two iOS devices but is more expensive.

For now we're going to leave Ganesh be and find a way (in the future) to only enable the pinning on Android in Graphite.

Bug: skia:14167
Change-Id: I8ef665ed35a7f572f19e296d8452190e28993367
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/654916
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>