Add SkSL test for indexing into a swizzle.
This is valid in ES2 but we didn't have any existing unit tests
which exercised this path. This actually caused an assertion in
SkRP, so it's now an "unsupported" path until proper support can
be added.

(I wouldn't be shocked if some GPUs fail the test as well; we can
disable them as necessary.)

Change-Id: Ide7c35a48496bec28798a9c8fe9d49d239fbbef5
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/655757
Reviewed-by: Arman Uguray <armansito@google.com>
Commit-Queue: Arman Uguray <armansito@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>