Rename SkImage::CompressionType -> SkTextureCompressionType
This appears to be in a few contexts outside of SkImages and this
will help with upcoming refactoring around the SkImage factories.

I inlined the one use of kCompressionTypeCount and decided to share
the value of kETC2_RGB8_UNORM with kETC1_RGB8 instead of
making a static constant variable.

go/skia-client-search showed no usage of the old
SkImage::CompressionType, only a few uses of SkImage::kETC1_SkCompressionType. Thus we only need an alias
for that static class variable until Chromium is migrated.

All other changes (the #include fixes) were a result of SkImage.h
not being included from GrTypesPriv.h. To assist with this,
I enforced IWYU on a few new files that gave me trouble as well as
having SkiaVulkan.h properly export the vulkan_core.h and friends
so IWYU doesn't try to make us directly include those.

Change-Id: I844372368478e3538fd4e8e01fdfefd988438e91
Bug: skia:13983
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/656237
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>