Fix indexed-swizzle lvalues in RP codegen.
The previous CL fixed non-lvalue swizzled-index-expressions. This CL
goes further and fixes lvalue swizzled-index-expressions; as such,
the fix from the previous CL is no longer needed.

This is pretty straightforward; we use the existing transform code
to rewrite the swizzle as before. We needed a place to stash the
rewritten expression in the lvalue, so that it wouldn't die before
its associated lvalue, so I added a field for it; not super elegant
but it seems pretty harmless in practice.

Change-Id: I949cf77067d6c81d52b76cd0f0114e9c0126d2ff
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/656250
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>