[graphite] Don't record draws that can't produce coverage
ClipStack checks the local bounds for empty shapes that collapse to a
point or a line. This could have been moved into Device, but all the
rest of the style impacts on draw bounds are part of
`applyClipToDraw`.

Also fixed a bug with inverse-fill clip checking. It was using a
device-space shape for checking against the clip elements but was still
using the original local-to-device transform.

Bug: b/266443584
Change-Id: Ia66c3497ecaa342edd3c477bb5edc4b05f65f9f0
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/657376
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>